% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleTechReps.R
\name{corr.plot}
\alias{corr.plot}
\title{Correlation between technical replicates}
\usage{
corr.plot(
  x,
  rep1,
  rep2,
  save = TRUE,
  filetype = "pdf",
  col = "red",
  dot.size = 0.5,
  text.size = 5,
  nrow = 4,
  ncol = 4,
  label.size = 1,
  dpi = 80
)
}
\description{
This function makes scatter plots to visualize the
correlation between a given pair of technical replicates (1 vs 2)
for each sample.
}
\author{
Chathurani Ranathunge
}
