% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_fit_df}
\alias{covid_fit_df}
\title{Suvarna et al 2021 LFQ data (fit object)}
\format{
An object of class "MArrayLM"
}
\usage{
data(covid_fit_df)
}
\description{
An object of class "MArrayLM" from running find_dep on covid_norm_df
}
\references{
\url{https://www.frontiersin.org/articles/10.3389/fphys.2021.652799/full#h3}
}
\keyword{datasets}
