% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_norm_df}
\alias{covid_norm_df}
\title{Suvarna et al 2021 LFQ data (normalized)}
\format{
A data frame with 230 rows (proteins) and 35 columns (samples)
}
\usage{
data(covid_norm_df)
}
\description{
A dataframe containing normalized LFQ protein intensity data for 230
proteins in 35 samples (a subset of the original data set)
}
\references{
\url{https://www.frontiersin.org/articles/10.3389/fphys.2021.652799/full#h3}
}
\keyword{datasets}
