% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDF.R
\name{create.df}
\alias{create.df}
\title{Create data frame with LFQ intensities}
\usage{
create.df(x, zeroNA = TRUE, logtr = TRUE, base = 2)
}
\description{
This function creates a data frame with LFQ intensities from the filtered
proteins from filter.prot function. Then converts zeros to NA.
}
\author{
Chathurani Ranathunge
}
