% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_fit_df}
\alias{ecoli_fit_df}
\title{Cox et al 2014 LFQ data (fit object)}
\format{
An object of class "MArrayLM"
}
\usage{
data(ecoli_fit_df)
}
\description{
An object of class "MArrayLM" from running find_dep on ecoli_norm_df
}
\references{
\url{https://europepmc.org/article/MED/24942700#id609082}
}
\keyword{datasets}
