% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_norm_df}
\alias{ecoli_norm_df}
\title{Cox et al 2014 LFQ data (normalized)}
\format{
A data frame with 4360 rows (proteins) and 6 columns (samples)
}
\usage{
data(ecoli_norm_df)
}
\description{
A dataframe containing normalized LFQ protein intensity data for 4360
proteins in 6 samples
}
\references{
\url{https://europepmc.org/article/MED/24942700#id609082}
}
\keyword{datasets}
