% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProteins.R
\name{filter.NA}
\alias{filter.NA}
\title{Filter missing data}
\usage{
filter.NA(x)
}
\description{
This function removes proteins (rows) that are absent (NA)
across all samples and samples (columns) with no protein data
}
\author{
Chathurani Ranathunge
}
