% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProteins.R
\name{filter.prot}
\alias{filter.prot}
\title{Filter proteins}
\usage{
filter.prot(x)
}
\description{
This function filters out proteins with a "+" sign for "Reverse",
"Only identified by site", and "Potential contaminants" columns in the
proteinGroups.txt file.
It also removes proteins identified with less than two unique peptides.
}
\author{
Chathurani Ranathunge
}
