% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProteins.R
\name{filterbygroup_na}
\alias{filterbygroup_na}
\title{Filter proteins by group level missing data}
\usage{
filterbygroup_na(df, set_na = 0.33)
}
\arguments{
\item{df}{A \code{raw_df} object (output of \code{\link{create_df}})}

\item{set_na}{The percentage of missing data allowed in any group.
Default is 0.33.}
}
\value{
A \code{raw_df} object or a data frame of filtered proteins
as rows and sample LFQ intensities as columns.
}
\description{
This function filters out proteins that exceed a given
missing data percentage in each group.
}
\details{
\itemize{\item This function first
extracts group or condition information from the \code{raw_df} object and
assigns samples to their groups.
\item It then removes proteins (rows) from the data frame if the percentage
of NAs in any one of the given groups exceeds the threshold indicated by
\code{set_na} (default is 0.33)}
}
\examples{
\dontrun{

## Missing data percentage allowed in each group = 0.33
raw_filtered <- filterbygroup_na(raw)

## Missing data percentage allowed in each group = 0.5
raw_filtered1 <- filterbygroup_na(raw, set_na = 0.5)
}
}
\author{
Chathurani Ranathunge
}
