% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDE.R
\name{find.DEP}
\alias{find.DEP}
\title{Identify differentially expressed proteins between groups}
\usage{
find.DEP(
  x,
  save.output = TRUE,
  filename = "LIMMA_output",
  save.tophits = TRUE,
  ntop = 20
)
}
\description{
This function identifies significantly differentially expressed
proteins between two groups with LIMMA
}
\author{
Chathurani Ranathunge
}
