% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDE.R
\name{find_dep}
\alias{find_dep}
\title{Identify differentially expressed proteins between groups}
\usage{
find_dep(
  norm_df,
  save_output = FALSE,
  save_tophits = FALSE,
  cutoff = 0.05,
  n_top = 20
)
}
\arguments{
\item{norm_df}{A \code{norm_df} object.}

\item{save_output}{Logical. If \code{TRUE} saves results from the
differential expression analysis in a text file labeled "limma_output.txt"
in the working directory.}

\item{save_tophits}{Logical. If \code{TRUE} saves \code{n_top}
number of top hits from the differential expression analysis in a text file
labeled "TopHits.txt" in the working directory.}

\item{cutoff}{Cutoff value for p-values and adjusted p-values. Default is
0.05.}

\item{n_top}{The number of top differentially expressed proteins to save in
the "TopHits.txt" file. Default is \code{20}.}
}
\value{
A \code{fit_df} object, which is similar to a \code{limma}
\code{fit} object.
}
\description{
This function performs differential expression analysis
on protein intensity data with limma.
}
\details{
\itemize{\item
It is important that the data is first log-transformed, ideally,
imputed, and normalized before performing differential expression analysis.
\item \code{save_output} saves the complete results table from the
differential expression analysis.
\item \code{save_tophits} first subsets the results to those with absolute
log fold change of more than 1, performs multiple correction with
"Benjamini Hochberg" method and outputs the top \code{n_top} results based
on lowest p-value and adjusted p-value.
\item If the number of hits with absolute log fold change of more than 1 is
less than \code{n_top}, \code{find_dep} prints only those with
log-fold change > 1 to "TopHits.txt"}
}
\examples{
\dontrun{

# Normalize an already imputed data set.
raw_nm <- normalize_data(raw_imp)

# Perform differential expression analysis.
fit <- find_dep(raw_nm)
}

}
\seealso{
\itemize{\item\code{normalize_data}
\item\code{\link[limma:lmFit]{lmFit}},
\code{\link[limma:eBayes]{eBayes}},
\code{\link[limma:topTable]{topTable}}, and
\code{\link[limma:write.fit]{write.fit}} functions from the
\code{\link[limma]{limma}} package.}
}
\author{
Chathurani Ranathunge
}
