% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDE.R
\name{heatmap.DE}
\alias{heatmap.DE}
\title{Heat map of differentially expressed proteins}
\usage{
heatmap.DE(
  x,
  y,
  cutoff = 0.05,
  FC = 1,
  adj.P = 0.05,
  sig = "adj.P",
  palette = "YlGnBu",
  save = TRUE,
  filename = "HeatmapDE",
  filetype = "pdf",
  dpi = 80,
  height = 7,
  width = 7
)
}
\description{
This function help visualize the results from the differential
expression analysis using a heat map.
}
\author{
Chathurani Ranathunge
}
