% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNA.R
\name{heatmap.NA}
\alias{heatmap.NA}
\title{Visualize missing data distribution}
\usage{
heatmap.NA(
  df,
  reorder.x = FALSE,
  reorder.y = FALSE,
  x.FUN = mean,
  y.FUN = mean,
  na.rm = T,
  col.na = "black",
  col.val = "grey",
  plot.title.size = 5,
  x.text.size = 2,
  y.text.size = 1,
  save = TRUE,
  filetype = "pdf",
  filename = "MissingData_heatmap",
  width = 15,
  height = 15,
  dpi = 80
)
}
\description{
This function visualizes the distribution of missing data using
a heatmap
}
\author{
Chathurani Ranathunge
}
