% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNA.R
\name{heatmap_na}
\alias{heatmap_na}
\title{Visualize missing data}
\usage{
heatmap_na(
  df,
  protein_range = "ALL",
  sample_range = "ALL",
  reorder_x = FALSE,
  reorder_y = FALSE,
  x_fun = mean,
  y_fun = mean,
  palette = "viridis",
  text_size = 10,
  save = FALSE,
  file_type = "pdf",
  file_name = "MissingData_heatmap",
  plot_width = 15,
  plot_height = 15,
  dpi = 80
)
}
\arguments{
\item{df}{A \code{raw_df} object (output from \code{\link{create_df}}).}

\item{protein_range}{The range of proteins to plot. Default is \code{ALL},
meaning all the proteins in the data frame.}

\item{sample_range}{The range of samples to plot. Default is \code{ALL},
meaning all the samples in the data frame.}

\item{reorder_x}{Logical. If \code{TRUE} samples on the x axis are reordered
using the function given in \code{x_fun}. Default is \code{FALSE}.}

\item{reorder_y}{Logical. If \code{TRUE} proteins in the y axis are reordered
using the function given in \code{y_fun}. Default is \code{FALSE}.}

\item{x_fun}{Function to reorder samples along the x axis. Possible options
include \code{mean} and \code{sum}. Default is \code{mean}.}

\item{y_fun}{Function to reorder proteins along the y axis. Possible options
include \code{mean} and \code{sum}. Default is \code{mean}.}

\item{palette}{Viridis color palette option for plots. Default is
\code{"viridis"}. See
\code{\link[viridisLite:viridis]{viridis}}
for available options.}

\item{text_size}{Text size for axis labels. Default is \code{10}.}

\item{save}{Logical. If \code{TRUE} saves a copy of the plot in the
working directory.}

\item{file_type}{File type to save the heatmap. Default is \code{"pdf"}.}

\item{file_name}{File name to save the heatmap.
Default is \code{"MissingData_heatmap"}.}

\item{plot_width}{Width of the plot. Default is \code{15}.}

\item{plot_height}{Height of the plot. Default is \code{15}.}

\item{dpi}{Plot resolution. Default is \code{80}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
This function visualizes the patterns of missing value
occurrence using a heatmap.
}
\details{
This function visualizes patterns of missing value occurrence using a
heatmap. User can choose to reorder the axes using the available functions
(\code{x_fun}, \code{y_fun}) to better understand the underlying cause of
missing data.
}
\examples{
\dontrun{
## Create a raw_df object from a proteinGroups.txt file.
raw <- create_df(prot.groups = "./proteinGroups.txt")

## Missing data heatmap with default settings.
heatmap_na(raw)

## Missing data heatmap with x and y axes reordered by sum of intensity.
heatmap_na(raw,
  reorder_x = TRUE, reorder_y = TRUE, x_fun = sum,
  y_fun = sum
)
}

}
\seealso{
\code{\link{create_df}}
}
\author{
Chathurani Ranathunge
}
