% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNA.R
\name{impute.densplot}
\alias{impute.densplot}
\title{Visualize the imputation effects}
\usage{
impute.densplot(
  original,
  imputed,
  global = TRUE,
  alpha = 0.25,
  lwd = 0.1,
  xlabel = "Intensity",
  ylabel = "Density",
  xlab.size = 7,
  ylab.size = 7,
  col1 = "blue",
  col2 = "red",
  strip.txt.size = 5,
  strip.ln.col = "grey",
  strip.ln.size = 0.5,
  nrow = 15,
  ncol = 5,
  dpi = 80,
  filename = "Impute_densityplot",
  filetype = "pdf",
  width = 7,
  height = 7,
  save = TRUE
)
}
\description{
This function compares imputed data to original data with a
user-defined plot.
}
\author{
Chathurani Ranathunge
}
