% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNA.R
\name{impute.NA}
\alias{impute.NA}
\title{Impute missing data}
\usage{
impute.NA(
  x,
  method = "minProb",
  tune.sigma = 1,
  q = 0.001,
  maxiter = 6,
  ntree = 20,
  verbose = TRUE,
  nPcs = 2,
  center = TRUE,
  imp_var = FALSE
)
}
\description{
This function imputes missing data using a user-specified
imputation method
}
\author{
Chathurani Ranathunge
}
