% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNA.R
\name{impute.plot}
\alias{impute.plot}
\title{Visualize the impact of imputation}
\usage{
impute.plot(
  original,
  imputed,
  global = TRUE,
  text.size = 10,
  orig.col = "blue",
  imp.col = "red",
  nrow,
  ncol,
  save = TRUE,
  file.name = "Impute_plot",
  file.type = "pdf",
  plot.width = 7,
  plot.height = 7,
  dpi = 80
)
}
\arguments{
\item{original}{A \code{raw.df} object (output of \code{\link{create.df}})
containing missing values.}

\item{imputed}{An \code{imp.df} object obtained from running \code{impute.NA}
on the same data frame provided as \code{original}.}

\item{global}{Logical. If \code{TRUE} ({default}), a global density plot is
produced. If \code{FALSE}, sample-wise density plots are produced.}

\item{text.size}{Text size for plot labels, axis labels etc. Default is
\code{10}.}

\item{orig.col}{Fill color for the \code{original} data set.\cr
Default is "Blue."}

\item{imp.col}{Fill color color for the \code{imputed} data set.\cr
Default is "Red."}

\item{nrow}{Required if \code{global = FALSE} to indicate the number of rows
to print the plots.}

\item{ncol}{Required if \code{global = FALSE} to indicate the number of
columns to print the plots.}

\item{save}{Logical. If \code{TRUE} (default) saves a copy of the plot in the
working directory.}

\item{file.name}{file.name File name to save the density plot/s.
Default is \code{"Impute_plot."}}

\item{file.type}{File type to save the density plot/s.
Default is \code{"pdf"}.}

\item{plot.width}{Width of the plot. Default is \code{7}.}

\item{plot.height}{Height of the plot. Default is \code{7}.}

\item{dpi}{Plot resolution. Default is \code{80}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
This function generates density plots to visualize the impact of
missing data imputation on the data.
}
\details{
\itemize{\item Given two data frames, one with missing values
(\code{raw.df} object) and the other, an imputed data frame
(\code{imp.df} object) of the same data set, \code{impute.plot}
generates global or sample-wise density plots to visualize the
impact of imputation on the data set.
\item Note, when sample-wise option is selected (\code{global = FALSE}),
\code{nrow} * \code{ncol} should match the number of samples in the
data frame.}
}
\examples{
\dontrun{
## Create a raw.df object from a proteinGroups.txt file.
raw <- create.df(file.path = "./proteinGroups.txt")

## Impute missing values in the data frame.
imp_raw <- impute.NA(raw)

## Visualize the impact of missing data imputation with a global density plot.
impute.plot(original = raw, imputed = imp_raw)

## Make sample-wise density plots for a data set of 25 samples.
impute.plot(raw, imp_raw, global = FALSE, nrow = 5, ncol = 5)

}
}
