% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNA.R
\name{impute_na}
\alias{impute_na}
\title{Impute missing values}
\usage{
impute_na(
  df,
  method = "minProb",
  tune_sigma = 1,
  q = 0.01,
  maxiter = 10,
  ntree = 20,
  n_pcs = 2
)
}
\arguments{
\item{df}{A \code{raw_df} object (output of \code{\link{create_df}})
containing missing values.}

\item{method}{Imputation method to use. Default is \code{"minProb"}.
Available methods: \code{"minDet", "RF", "kNN", and "SVD"}.}

\item{tune_sigma}{A scalar used in the \code{"minProb"} method for
controlling the standard deviation of the Gaussian distribution
from which random values are drawn for imputation.\cr
Default is 1.}

\item{q}{A scalar used in \code{"minProb"} and \code{"minDet"} methods
to obtain a low intensity value for imputation. \code{q} should be set to a
very low value. Default is 0.01.}

\item{maxiter}{Maximum number of iterations to be performed when using the
\code{"RF"} method. Default is \code{10}.}

\item{ntree}{Number of trees to grow in each forest when using the
\code{"RF"} method. Default is \code{20}.}

\item{n_pcs}{Number of principal components to calculate when using the
\code{"SVD"} method. Default is 2.}
}
\value{
An \code{imp.df} object, which is a data frame of intensities with no
missing values.
}
\description{
This function imputes missing values using a user-specified
imputation method.
}
\details{
\code{impute_na} function imputes missing values using a
user-specified imputation method from the available options, \code{minProb},
\code{minDet}, \code{kNN}, \code{RF}, and \code{SVD}.
}
\examples{
\dontrun{
## Create a raw_df object from a proteinGroups.txt file.
raw <- create_df(prot_groups = "./proteinGroups.txt")

## Impute missing values in the data frame using the default minProb
## method.
imp_raw <- impute_na(raw)

## Using the RF method with the number of iterations set at 5
## and number of trees set at 100.
imp_raw <- impute_na(raw, method = "RF", maxiter = 5, ntree = 100)

## Using the kNN method.
imp_raw <- impute_na(raw, method = "kNN")

## Using the SVD method with n_pcs set to 3.
imp_raw <- impute_na(raw, method = "SVD", n_pcs = 3)

## Using the minDet method with q set at 0.001.
imp_raw <- impute_na(raw, method = "minDet", q = 0.001)
}

}
\seealso{
More information on the available imputation methods can be found
in their respective packages.
\itemize{\item \code{\link{create_df}}
\item For \code{minProb} and
\code{minDet} methods, see
\code{imputeLCMD} package.
\item For Random Forest (\code{RF}) method, see
\code{\link[missForest]{missForest}}.
\item For \code{kNN} method, see \code{\link[VIM]{kNN}} from the
\code{\link[VIM]{VIM}} package.
\item For \code{SVD} method, see \code{\link[pcaMethods]{pca}} from the
\code{\link[pcaMethods]{pcaMethods}} package.}
}
\author{
Chathurani Ranathunge
}
