% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleNA.R
\name{impute_plot}
\alias{impute_plot}
\title{Visualize the impact of imputation}
\usage{
impute_plot(
  original,
  imputed,
  global = TRUE,
  text_size = 10,
  palette = "viridis",
  nrow,
  ncol,
  save = FALSE,
  file_name = "Impute_plot",
  file_type = "pdf",
  plot_width = 7,
  plot_height = 7,
  dpi = 80
)
}
\arguments{
\item{original}{A \code{raw_df} object (output of \code{\link{create_df}})
containing missing values.}

\item{imputed}{An \code{imp.df} object obtained from running \code{impute_na}
on the same data frame provided as \code{original}.}

\item{global}{Logical. If \code{TRUE} ({default}), a global density plot is
produced. If \code{FALSE}, sample-wise density plots are produced.}

\item{text_size}{Text size for plot labels, axis labels etc. Default is
\code{10}.}

\item{palette}{Viridis color palette option for plots. Default is
\code{"viridis"}. See
\code{\link[viridis: scale_color_viridis]{scale_color_viridis}}
for available options.}

\item{nrow}{Required if \code{global = FALSE} to indicate the number of rows
to print the plots.}

\item{ncol}{Required if \code{global = FALSE} to indicate the number of
columns to print the plots.}

\item{save}{Logical. If \code{TRUE} saves a copy of the plot in the
working directory.}

\item{file_name}{file_name File name to save the density plot/s.
Default is \code{"Impute_plot."}}

\item{file_type}{File type to save the density plot/s.
Default is \code{"pdf"}.}

\item{plot_width}{Width of the plot. Default is \code{7}.}

\item{plot_height}{Height of the plot. Default is \code{7}.}

\item{dpi}{Plot resolution. Default is \code{80}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
This function generates density plots to visualize the impact of
missing data imputation on the data.
}
\details{
\itemize{\item Given two data frames, one with missing values
(\code{raw_df} object) and the other, an imputed data frame
(\code{imp.df} object) of the same data set, \code{impute_plot}
generates global or sample-wise density plots to visualize the
impact of imputation on the data set.
\item Note, when sample-wise option is selected (\code{global = FALSE}),
\code{nrow} * \code{ncol} should match the number of samples in the
data frame.}
}
\examples{
\dontrun{
## Create a raw_df object from a proteinGroups.txt file.
raw <- create_df(prot.groups = "./proteinGroups.txt")

## Impute missing values in the data frame.
imp_raw <- impute_na(raw)

## Visualize the impact of missing data imputation with a global density
## plot.
impute_plot(original = raw, imputed = imp_raw)

## Make sample-wise density plots for a data set of 25 samples.
impute_plot(raw, imp_raw, global = FALSE, nrow = 5, ncol = 5)
}
}
\author{
Chathurani Ranathunge
}
