% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm.plot}
\alias{norm.plot}
\title{Visualize the effect of normalization}
\usage{
norm.plot(
  original,
  normalized,
  type = "box",
  text.size = 10,
  palette = "YlGnBu",
  save = TRUE,
  file.name = "Norm_plot",
  file.type = "pdf",
  dpi = 80,
  plot.width = 7,
  plot.height = 7
)
}
\arguments{
\item{original}{A \code{raw.df} object (output of \code{\link{create.df}})
containing missing values or an \code{imp.df} object after imputing the
missing values with \code{impute.NA}.}

\item{normalized}{A \code{norm.df} object after normalizing the data frame
provided as \code{original}.}

\item{type}{Type of plot to generate. Choices are "box" or "density." Default
is \code{"box."}}

\item{text.size}{Text size for plot labels, axis labels etc. Default is
\code{10}.}

\item{palette}{Color palette for box plots. Default is \code{"YlGnBu."}}

\item{save}{Logical. If \code{TRUE} (default) saves a copy of the plot in the
working directory.}

\item{file.name}{file.name File name to save the plot.
Default is \code{"Norm_plot."}}

\item{file.type}{File type to save the plot.
Default is \code{"pdf"}.}

\item{dpi}{Plot resolution. Default is \code{80}.}

\item{plot.width}{Width of the plot. Default is \code{7}.}

\item{plot.height}{Height of the plot. Default is \code{7}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
This function visualizes the impact of normalization on
the data with plots.
}
\details{
Given two data frames, one with data prior to normalization
(\code{original}) and one after normalization (\code{normalized}),
\code{norm.plot} generates side-by-side plots to visualize the effect of
normalization on the intensity data.
}
\examples{
\dontrun{
## Create a raw.df object from a proteinGroups.txt file.
raw <- create.df(file.path = "./proteinGroups.txt", log.tr = TRUE)

## Normalize an imputed data set.
raw_imp <- impute.NA(raw)
raw_nm <- normalize.data(raw_imp)

## Visualize normalization with box plots.
norm.plot(raw_imp, raw_nm)

## Visualize normalization with density plots.
norm.plot(raw_imp, raw_nm, type = "density")
}

}
\seealso{
\itemize{\item \code{\link{normalize.data}}
\item \code{create.df}
\item \code{impute.NA}
}
}
\author{
Chathurani Ranathunge
}
