% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize.data}
\alias{normalize.data}
\title{Normalize intensities}
\usage{
normalize.data(x, method = "quantile")
}
\description{
This function normalizes data using a user-specified
normalization method.
}
\author{
Chathurani Ranathunge
}
