% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize intensity data}
\usage{
normalize_data(df, method = "quantile")
}
\arguments{
\item{df}{A \code{raw.df} object, or ideally, an \code{imp.df} object with
missing values imputed.}

\item{method}{Name of the normalization method to use. Choices are
\code{"none", "scale", "quantile" or "cyclicloess."}
Default is \code{"quantile."}}
}
\value{
A \code{norm.df} object, which is a data frame with
normalized intensities.
}
\description{
This function normalizes data using a user-specified
normalization method.
}
\details{
\code{normalize_data} normalizes intensity values to achieve
consistency among samples. The function assumes that the intensities in the
data frame have been log-transformed, therefore, it is important to make sure
that \code{create_df} was run with \code{log.tr = TRUE} when creating the
\code{raw.df} object.
}
\examples{
\dontrun{
## Create a raw.df object from a proteinGroups.txt file.
raw <- create_df(file.path = "./proteinGroups.txt", log.tr = TRUE)

## Normalize a data set.
raw_nm <- normalize_data(raw, method = "cyclicloess")

## Normalize an imputed data set.
raw_imp <- impute_na(raw)
raw_nm <- normalize_data(raw_imp)

}


}
\seealso{
\itemize{\item \code{create_df}
\item \code{impute_na}
\item See \code{\link[limma]{normalizeBetweenArrays}} in the R package
\code{limma} for more information on the different normalization methods
available.}
}
\author{
Chathurani Ranathunge
}
