% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProteins.R
\name{onegroup.only}
\alias{onegroup.only}
\title{Proteins that are only expressed in a given group}
\usage{
onegroup.only(df, abs.group, pres.group, set.na = 0.33, save = TRUE)
}
\arguments{
\item{df}{A \code{raw.df} object (output of \code{\link{create.df}})}

\item{abs.group}{Name of the group in which proteins are not expressed.}

\item{pres.group}{Name of the group in which proteins are expressed.}

\item{set.na}{The percentage of missing data allowed in \code{pres.group}.
Default is 0.33.}

\item{save}{Logical. If \code{TRUE} (default), it saves the output in a text
file named "Group_\code{pres.group}_only.txt."}
}
\value{
A list of majority protein IDs.
}
\description{
This function outputs a list of proteins that are only
expressed (present) in one user-specified group while not expressed
(completely absent) in another user-specified group.
}
\details{
This function assumes that column headers in the \code{raw.df}
object provided as \code{df} follow "Group_UniqueSampleID" notation.
\itemize{\item Given a pair of groups, \code{onegroup.only} function
finds proteins that are only expressed in \code{pres.group} while
completely absent or not expressed in \code{abs.group}.}
}
\examples{
\dontrun{
## Create a raw.df object from a proteinGroups.txt file.
raw <- create.df(file.path = "./proteinGroups.txt")

## Save a list of proteins only expressed in group B, but absent in group A.
onegroup.only(raw, abs.group = "A", pres.group = "B")

## Save the above list in a variable.
protein_list <- onegroup.only(raw, abs.group = "A", pres.group = "B")
}
}
