% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlots.R
\name{predictor_plot}
\alias{predictor_plot}
\title{Visualize predictor (protein) variation among conditions}
\usage{
predictor_plot(
  df,
  type = "box",
  text.size = 10,
  palette = "YlGnBu",
  nrow,
  ncol,
  save = FALSE,
  file.name = "Predictor_plot",
  file.type = "pdf",
  dpi = 80,
  plot.width = 7,
  plot.height = 7
)
}
\arguments{
\item{df}{A \code{model.df} object from performing \code{pre_process}.}

\item{type}{Type of plot to generate. Choices are "box" or "density." Default
is \code{"box."}}

\item{text.size}{Text size for plot labels, axis labels etc. Default is
\code{10}.}

\item{palette}{Color palette for box plots. Default is \code{"YlGnBu."}}

\item{nrow}{The number of rows to print the plots.}

\item{ncol}{The number of columns to print the plots.}

\item{save}{Logical. If \code{TRUE} saves a copy of the plot in the
working directory.}

\item{file.name}{file.name File name to save the plot.
Default is \code{"Predictor_plot."}}

\item{file.type}{File type to save the plot.
Default is \code{"pdf"}.}

\item{dpi}{Plot resolution. Default is \code{80}.}

\item{plot.width}{Width of the plot. Default is \code{7}.}

\item{plot.height}{Height of the plot. Default is \code{7}.}
}
\value{
A \code{ggplot2} object
}
\description{
This function visualizes protein intensity differences among
conditions with plots.
}
\details{
This function visualizes condition-wise differences in protein
intensity using boxplots and/or density plots.
}
\examples{
\dontrun{

 ## Box plots
 predictor_plot(model_df, type = "box", nrow = 2, ncol = 4)

 ## Density plots
 predictor_plot(model_df, type = "density", nrow = 2, ncol = 4)

 }
}
\seealso{
\itemize{
\item \code{pre_process}, \code{rem_predictors}}
}
\author{
Chathurani Ranathunge
}
