% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleTechReps.R
\name{rem.sample}
\alias{rem.sample}
\title{Remove user-specified samples}
\usage{
rem.sample(x, rem = "")
}
\description{
This function removes user-specified samples (columns)
from the database
}
\author{
Chathurani Ranathunge
}
