% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelProteins.R
\name{rem_feature}
\alias{rem_feature}
\title{Remove user-specified proteins (predictors) from the data frame}
\usage{
rem_feature(model_df, rem_protein)
}
\arguments{
\item{model_df}{A \code{model_df} object.}

\item{rem_protein}{Name of the protein to remove.}
}
\value{
A \code{model_df} object.
}
\description{
This function removes user-specified proteins from the
data frame.
}
\details{
\itemize{\item After visualizing protein intensity variation
among conditions with \code{predictor_plot}, you can choose to remove
specific proteins (predictors) from the data frame before modeling if they
do not show distinct patterns of variation among conditions. For
example, these proteins could have overlapping density distributions.
}
}
\examples{
\dontrun{

## Remove protein "A113403" from model_df data frame.
model_df <- rem_feature(model_df, "A113403")
}

}
\seealso{
\code{\link{predictor_plot}}, \code{\link{pre_process}}
}
\author{
Chathurani Ranathunge
}
