% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelProteins.R
\name{test_models}
\alias{test_models}
\title{Test machine learning models on test data}
\usage{
test_models(
  model.list,
  split.df,
  type = "prob",
  save.confusionmatrix = FALSE,
  ...
)
}
\arguments{
\item{model.list}{A \code{model.list} object from performing
\code{train_models}.}

\item{split.df}{A \code{split.df} object from performing \code{split_df}.}

\item{type}{Type of output. Set \code{type} as "prob" (default) to output
class probabilities, and "raw" to output class predictions.}

\item{save.confusionmatrix}{Logical. If \code{TRUE}, a tab-delimited
text file ("Confusion_matrices.txt") with confusion matrices in the
long-form data format will be saved in the working directory.
See below for more details.}

\item{...}{Additional arguments to be passed on to
\code{\link[stats:predict]{predict}}.}
}
\value{
\itemize{\item \code{probability.list}: If \code{type = "prob"}, a list of
data frames containing class probabilities for each method in the
\code{model.list} will be returned.
\item \code{prediction.list}: If \code{type = "raw"}, a list of factors
containing class predictions for each method will be returned.}
}
\description{
This function can be used to test models generated using
different machine learning algorithms on a test data set
}
\details{
\itemize{\item \code{test_models} function uses
models obtained from \code{train_models} to predict a given test data set.
\item Setting \code{type = "raw"} is required to obtain confusion matrices.
\item Setting \code{type = "prob"} (default) will output a list of
probabilities that can be used to generate ROC curves with \code{ROC_plot}.}
}
\examples{
\dontrun{
#Test a list of models on a test data set and output class probabilities,
model_prob <- test_models(model.list = model_fit, split.df = split_df)

#Save confusion matrices and output class predictions
model_pred <- test_models(model.list = model_fit,
                         split.df = split_df,
                         type = "raw",
                         save.confusionmatrix = TRUE)
 }
}
\seealso{
\itemize{
\item \code{split_df}
\item \code{train_models}
\item \code{\link[stats:predict]{stats: predict}}
\item \code{\link[caret:confusionMatrix]{caret: confusionMatrix}}
}
}
\author{
Chathurani Ranathunge
}
