% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelProteins.R
\name{train_model}
\alias{train_model}
\title{Train machine learning models on protein intensity data}
\usage{
train_model(
  split.df,
  resample.method = "repeatedcv",
  resample.iterations = 10,
  num.repeats = 3,
  algorithm.list,
  ...
)
}
\arguments{
\item{split.df}{A \code{split.df} object from performing \code{split_df}.}

\item{resample.method}{The resampling method to use. Default is "repeatedcv"
for repeated cross validation.
See \code{\link[caret:trainControl]{trainControl}} for
details on other available methods.}

\item{resample.iterations}{Number of resampling iterations. Default is 10.}

\item{num.repeats}{The number of complete sets of folds to compute (For
\code{resampling method = "repeatedcv"} only).}

\item{algorithm.list}{A list of classification or regression algorithms to use.
A full list of machine learning algorithms available through
the \code{\link[caret]{caret package}} can be found here:
\url{http://topepo.github.io/caret/train-models-by-tag.html}. see below for
default options.}
}
\value{
A list of class \code{train} for each machine-learning algorithm.
See \code{\link[caret:train]{train}} for more information on accessing
different elements of this list.
}
\description{
This function can be used to train models on protein intensity
data using a number of different machine learning algorithms
}
\details{
\itemize{\item \code{train_model} function can be used to first
define the control parameters to be used in training models, calculate
resampling-based performance measures for models based on a given set of
machine-learning algorithms, and output the best model for each algorithm.
\item In the event that \code{algorithm.list} is not provided, a default
list of five classification-based machine-learning algorithms will be used
for building and training models. Default \code{algorithm.list}:
"rf", "ranger", "earth", "knn", "svmLinear."
\item Note: Models that failed to build are removed from the output. }
}
\examples{
\dontrun{
#Fit models based on the default list of ML algorithms.
model_fit <- train_model(split.df = split_df)

#Fit models using a user-specified list of ML algorithms.
model_fit <- train_model(split.df = split_df,
                         algorithm.list = c("svmRadial", "rpart"))

#Change resampling method and resampling iterations.
model_fit <- train_model(split.df = split_df,
                         resample.method = "cv",
                         resample.iterations = 50)
 }


}
\seealso{
\itemize{
\item \code{pre_process}
\item \code{\link[caret:trainControl]{caret: trainControl}}
\item \code{\link[caret:train]{caret: train}}
}
}
\author{
Chathurani Ranathunge
}
