% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDE.R
\name{volcano.plot}
\alias{volcano.plot}
\title{Volcano plot}
\usage{
volcano.plot(
  x,
  adjust.method = "BH",
  cutoff = 0.05,
  FC = 1,
  line.FC = TRUE,
  line.P = TRUE,
  filename = "Volcano_plot",
  filetype = "pdf",
  height = 2,
  width = 2,
  col.by = "adjP",
  nsig.col = "black",
  sig.col = "red",
  ln.col = "red",
  label.top = FALSE,
  label.n = 10,
  dpi = 80,
  save = FALSE
)
}
\description{
This function help visualize the results from the differential
expression analysis using a volcano plot.
}
\author{
Chathurani Ranathunge
}
