% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleTechReps.R
\name{aver.techreps}
\alias{aver.techreps}
\title{Compute average intensity}
\usage{
aver.techreps(x)
}
\description{
This function computes average intensities across
technical replicates for each sample.
}
\author{
Chathurani Ranathunge
}
