% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProteins.R
\name{filter.NA}
\alias{filter.NA}
\title{Filter out proteins and/ or samples that do not have valid data}
\usage{
filter.NA(df)
}
\arguments{
\item{df}{A \code{raw.df} object (output of \code{\link {create.df}}).}
}
\value{
A \code{raw.df} object or a data frame with no empty (all NAs) rows
or columns.
}
\description{
This function removes proteins (rows) that have missing data
(NAs) across all samples and samples (columns) with missing data across all
proteins
}
\details{
\itemize{\item This function first removes proteins (rows) with
missing values (NAs) across all samples.
\item Then if present, it removes samples (columns) with missing data
across all proteins }
}
\examples{
\dontrun{
raw <- create.df(file.path = "./proteinGroups.txt")
raw_1 <- filter.NA(raw)
}
}
