% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProteins.R
\name{filterby.groupNA}
\alias{filterby.groupNA}
\title{Filter proteins by group level missing data}
\usage{
filterby.groupNA(x, set.na = 0.33)
}
\description{
This function filters out proteins with more than 33\% missing
data in each group.
}
\author{
Chathurani Ranathunge
}
