% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProteins.R
\name{filterbygroup_na}
\alias{filterbygroup_na}
\title{Filter proteins by group level missing data}
\usage{
filterbygroup_na(df, set_na = 0.33, filter_condition = "either")
}
\arguments{
\item{set_na}{The proportion of missing data allowed.
Default is 0.33.}

\item{filter_condition}{If set to \code{"each"}, proteins that exceed
the missing value proportion threshold set by \code{set_na} in each group
will be removed.
If set to \code{"either"}(default), proteins that exceed the missing value
proportion threshold set by \code{set_na} in at least one group will be
removed.}

\item{raw_df}{A \code{raw_df} object (output of \code{\link{create_df}})}
}
\value{
A \code{raw_df} object.
}
\description{
This function filters out proteins that exceed a given
missing data percentage in each group.
}
\details{
\itemize{\item This function first
extracts group or condition information from the \code{raw_df} object and
assigns samples to their groups.
\item If \code{filter_condition = "each"}, it then removes proteins (rows)
from the data frame if the proportion of NAs in each group exceeds the
threshold indicated by \code{set_na} (default is 0.33)}. This option is
more lenient in comparison to \code{filter_condition = "either"}, where
proteins that exceeds the missing data threshold in either group gets
removed from the data frame.
}
\examples{
\dontrun{

## Missing data percentage allowed in each group = 0.33
raw_filtered <- filterbygroup_na(raw)

## Missing data percentage allowed in each group = 0.5
raw_filtered1 <- filterbygroup_na(raw, set_na = 0.5)
}
}
\author{
Chathurani Ranathunge
}
