% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDE.R
\name{find.DEP}
\alias{find.DEP}
\title{Identify differentially expressed proteins between groups}
\usage{
find.DEP(df, save.output = TRUE, save.tophits = TRUE, n.top = 20)
}
\arguments{
\item{df}{A \code{norm.df} object.}

\item{save.output}{Logical. If \code{TRUE} (default) saves results from the
differential expression analysis in a text file labeled "limma_output.txt"
in the working directory.}

\item{save.tophits}{Logical. If \code{TRUE} (default) saves \code{n.top}
number of top hits from the differential expression analysis in a text file
labeled "TopHits.txt" in the working directory.}

\item{n.top}{The number of top differentially expressed proteins to save in
the "TopHits.txt" file. Default is \code{20}.}
}
\value{
A \code{fit.df} object, which is similar to a \code{limma}
\code{fit} object.
}
\description{
This function performs differential expression analysis
on protein intensity data with limma.
}
\details{
\itemize{\item
It is important that the data is first log-transformed, ideally,
imputed, and normalized before performing differential expression analysis.
\item \code{save.output} saves the complete results table from the
differential expression analysis.
\item \code{save.tophits} first subsets the results to those with absolute
log fold change of more than 1, performs multiple correction with
"Benjamini Hochberg" method and outputs the top \code{n.top} results based
on lowest p-value and adjusted p-value.
\item If the number of hits with absolute log fold change of more than 1 is
less than \code{n.top}, \code{find.DEP} prints only those with
log-fold change > 1 to "TopHits.txt"}
}
\examples{
\dontrun{

#Normalize an already imputed data set.
raw_nm <- normalize.data(raw_imp)

#Perform differential expression analysis.
fit <- find.DEP(raw_nm)

}

}
\seealso{
\itemize{\item\code{normalize.data}
\item\code{\link[limma:lmFit]{lmFit}},
\code{\link[limma:eBayes]{eBayes}},
\code{\link[limma:topTable]{topTable}}, and
\code{\link[limma:write.fit]{write.fit}} functions from the
\code{\link[limma]{limma}} package.}
}
\author{
Chathurani Ranathunge
}
