% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{load.data}
\alias{load.data}
\title{Load input data}
\usage{
load.data(file.path = "")
}
\description{
This function loads the proteinGroups.txt file produced by
MaxQuant.
}
\author{
Chathurani Ranathunge
}
