% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm.plot}
\alias{norm.plot}
\title{Visualize the effect of normaliztion on the data}
\usage{
norm.plot(
  normalized,
  original,
  type = "box",
  alpha = 0.25,
  lwd = 0.1,
  xlabel = "",
  ylabel = "",
  xlab.size = 7,
  ylab.size = 7,
  palette = "YlGnBu",
  strip.txt.size = 5,
  strip.ln.size = 0.5,
  lab.text.size = 5,
  save = FALSE,
  filename = "Norm_plot",
  filetype = "pdf",
  dpi = 80,
  width = 7,
  height = 7
)
}
\description{
This function helps visualize the impact of normalization with
box plots or density plots
}
\author{
Chathurani Ranathunge
}
