% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProteins.R
\name{onegroup.only}
\alias{onegroup.only}
\title{Proteins only found in one group}
\usage{
onegroup.only(x, abs.group = "", pres.group = "", set.na = 0.33)
}
\description{
This function outputs a list of proteins absent in one group
and present in more than 66\% of the samples in the other group.
}
\details{
Comparisons are limited to two groups.
}
\author{
Chathurani Ranathunge
}
