% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlots.R
\name{performance_plot}
\alias{performance_plot}
\title{Model performance plot}
\usage{
performance_plot(
  model_list,
  type = "box",
  text_size = 10,
  palette = "viridis",
  save = FALSE,
  file_name = "Performance_plot",
  file_type = "pdf",
  plot_width = 7,
  plot_height = 7,
  dpi = 80
)
}
\arguments{
\item{model_list}{A \code{model_list} object from performing
\code{train_models}.}

\item{type}{Type of plot to generate. Choices are "box" or "dot."
Default is \code{"box."} for boxplots.}

\item{text_size}{Text size for plot labels, axis labels etc. Default is
\code{10}.}

\item{palette}{Viridis color palette option for plots. Default is
\code{"viridis"}. See
\code{\link[viridis: scale_color_viridis]{scale_color_viridis}}
for available options.}

\item{save}{Logical. If \code{TRUE} saves a copy of the plot in the
working directory.}

\item{file_name}{file_name File name to save the plot.
Default is \code{"Performance_plot."}}

\item{file_type}{File type to save the plot.
Default is \code{"pdf"}.}

\item{plot_width}{Width of the plot. Default is \code{7}.}

\item{plot_height}{Height of the plot. Default is \code{7}.}

\item{dpi}{Plot resolution. Default is \code{80}.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function visualizes model performance
}
\details{
\itemize{\item \code{performance_plot} uses resampling results from
models included in the \code{model_list} to generate plots of model
performance.
\item The default metrics used for classification based models are "Accuracy"
and "Kappa."
\item These metric types can be changed by providing additional arguments to
the \code{train_models} function. See \code{\link[caret: train]{train}} and
\code{\link[caret: trnControl]{trnControl}} for more information.}
}
\examples{
\dontrun{

## Train models using default settings
model_list <- train_models(split_df)

## Generate box plots to visualize performance of different methods
performance_plot(model_list)

## Generate dot plots
performance_plot(model_list, type = "dot")
}
}
\seealso{
\itemize{
\item \code{train_models}
\item\code{\link[caret: resample]{caret: resamples}}
\item \code{\link[caret:train]{caret: train}}
\item \code{\link[caret:trnControl]{caret: trnControl}}
}
}
\author{
Chathurani Ranathunge
}
