% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelProteins.R
\name{pre_process}
\alias{pre_process}
\title{Pre-process protein intensity data for modeling}
\usage{
pre_process(
  fit.df,
  norm.df,
  sig = "adjP",
  sig.cutoff = 0.05,
  FC = 1,
  n.top = 20,
  find.highcorr = TRUE,
  corr.cutoff = 0.9,
  save.corrmatrix = FALSE,
  rem.highcorr = TRUE
)
}
\arguments{
\item{fit.df}{A \code{fit.df} object from performing \code{find_dep}.}

\item{norm.df}{The \code{norm.df} object from which the \code{fit.df} object
was obtained.}

\item{sig}{Criteria to denote significance in differential expression.
Choices are \code{"adjP"} (default) for adjusted p-value or \code{"P"}
for p-value.}

\item{sig.cutoff}{Cutoff value for p-values and adjusted p-values in
differential expression. Default is 0.05.}

\item{FC}{Minimum absolute log-fold change to use as threshold for
differential expression. Default is 1.}

\item{n.top}{The number of top hits from \code{find_dep} to be used in
modeling. Default is 20.}

\item{find.highcorr}{Logical. If \code{TRUE} (default), finds highly
correlated proteins.}

\item{corr.cutoff}{A numeric value specifying the correlation cutoff.
Default is 0.90.}

\item{save.corrmatrix}{Logical. If \code{TRUE}, saves a copy of the
protein correlation matrix in a tab-delimited text file labeled
"Protein_correlation.txt" in the working directory.}

\item{rem.highcorr}{Logical. If \code{TRUE} (default), removes highly
correlated proteins (predictors or features).}
}
\value{
A \code{model.df} object, which is a data frame of protein
intensities with proteins indicated by columns.
}
\description{
This function pre-processes protein intensity data from
the top differentially expressed proteins identified with \code{find_dep} for
modeling.
}
\details{
This function creates a data frame that contains protein intensities
for a user-specified number of top differentially expressed proteins.
\itemize{\item Using \code{find.highcorr = TRUE}, highly correlated
proteins can be identified, and can be removed with
\code{rem.highcorr = TRUE}.
\item Note: Most models will benefit from reducing correlation between
proteins (predictors or features), therefore we recommend removing those
proteins at this stage to reduce pairwise-correlation.
\item If no proteins meet the significance threshold for differential
expression, you may adjust \code{sig}, \code{FC}, and \code{sig.cutoff}
accordingly to obtain a set of proteins for modeling.}
}
\examples{
\dontrun{

 ## Create a model.df object with default settings.
 model_df <- pre_process(fit_df, norm_df)

 ## Change the correlation cutoff.
 model_df <- pre_process(fit_df, norm_df, corr.cutoff = 0.95)
 }


}
\seealso{
\itemize{
\item \code{find_dep}, \code{normalize_data}
\item \code{\link[caret:findCorrelation]{caret: findCorrelation}}}
}
\author{
Chathurani Ranathunge
}
