% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleTechReps.R
\name{rem_sample}
\alias{rem_sample}
\title{Remove user-specified samples}
\usage{
rem_sample(df, rem)
}
\arguments{
\item{df}{A \code{raw_df} object.}

\item{rem}{Name of the sample to remove.}
}
\value{
A \code{raw_df} object.
}
\description{
This function removes user-specified samples from the
data frame.
}
\details{
\itemize{\item\code{rem_sample} assumes that sample names follow the
"Group_Sample_TechnicalReplicate" notation.
\item If all the technical replicates representing a sample needs to be
removed, provide "Group_UniqueSampleID" as \code{rem}.
\item If a specific technical replicate needs to be removed in case it
shows weak correlation with other technical replicates for example, you can
remove that particular technical replicate by providing
"Group_UniqueSampleID_TechnicalReplicate" as \code{rem}.
}
}
\examples{
\dontrun{

## Create a raw_df object from a proteinGroups.txt file.
raw <- create_df(file.path = "./proteinGroups.txt", log.tr = TRUE)

## Remove all technical replicates of "A_10"
raw_1 <- rem_sample(raw, "A_10")

## Remove only technical replicate number 2 of "A_10"
raw_1 <- rem_sample(raw, "A_10_2")
}

}
\seealso{
\code{\link{corr_plot}}, \code{\link{create_df}}
}
\author{
Chathurani Ranathunge
}
