% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlots.R
\name{roc_plot}
\alias{roc_plot}
\title{ROC plot}
\usage{
roc_plot(
  probability_list,
  split_df,
  ...,
  multiple_plots = TRUE,
  text_size = 10,
  palette = "viridis",
  save = FALSE,
  file_name = "ROC_plot",
  file_type = "pdf",
  plot_width = 7,
  plot_height = 7,
  dpi = 80
)
}
\arguments{
\item{probability_list}{A \code{probability_list} object from performing
\code{test_models} with \code{type = "prob"}.}

\item{split_df}{A \code{split_df} object from performing \code{split_data}}

\item{...}{Additional arguments to be passed on to
\code{\link[pROC: roc]{roc}}.}

\item{multiple_plots}{Logical. If \code{FALSE} plots all ROC curves
representing algorithms included in the \code{probability_list} in a single
plot.}

\item{text_size}{Text size for plot labels, axis labels etc. Default is
\code{10}.}

\item{palette}{Viridis color palette option for plots. Default is
\code{"viridis"}. See
\code{\link[viridis: scale_color_viridis]{scale_color_viridis}}
for available options.}

\item{save}{Logical. If \code{TRUE} saves a copy of the plot in the
working directory.}

\item{file_name}{file_name File name to save the plot.
Default is \code{"ROC_plot."}}

\item{file_type}{File type to save the plot.
Default is \code{"pdf"}.}

\item{plot_width}{Width of the plot. Default is \code{7}.}

\item{plot_height}{Height of the plot. Default is \code{7}.}

\item{dpi}{Plot resolution. Default is \code{80}.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function generates Receiver Operator Characteristic (ROC)
curves to evaluate models
}
\details{
\itemize{\item \code{roc_plot} first uses probabilities generated
during \code{test_models} to build a ROC object.
\item Next, relevant information is extracted from the ROC object to
plot the ROC curves.}
}
\examples{
\dontrun{

## Test models
model_probs <- test_models(
  model_list = model_fit, split_df = split_df,
  type = "prob"
)

## Plot all ROC curves in one plot
roc_plot(model_probs, multiple_plots = FALSE)

## Plot ROC curves separately for each algorithm
roc_plot(model_probs)
}
}
\seealso{
\itemize{
\item \code{test_models}
\item\code{\link[pROC: roc]{pROC: roc}}
}
}
\author{
Chathurani Ranathunge
}
